FEAT: Se reordenaron los metodos de Add*FromAssemblyContaining a su proyecto correspondiente, junto con darles un mejor nombre
