+++ VSlices.Core.Abstracts/Extensions/ServiceCollectionExtensions.cs
@@ -56,7 +56,7 @@ public static class ServiceCollectionExtensions
     /// <param name="services">Service collection</param>
     /// <returns>Service collection</returns>
     /// <exception cref="InvalidOperationException"></exception>
-    public static IServiceCollection AddCoreDependenciesFrom<TAnchor>(this IServiceCollection services)
+    public static IServiceCollection AddCoreDependenciesFromAssemblyContaining<TAnchor>(this IServiceCollection services)
     {
         var definerTypes = typeof(TAnchor).Assembly.ExportedTypes
             .Where(e => typeof(IUseCaseDependencyDefinition).IsAssignableFrom(e))
@@ -72,6 +72,31 @@ public static class ServiceCollectionExtensions
             }
 
             defineDependenciesMethod.Invoke(null, new object?[] { services });
+        }
+
+        return services;
+    }
+
+    /// <summary>
+    /// Adds <see cref="IHandler{TRequest,TResponse}"/> implementations from the specified assembly of the <typeparamref name="TAnchor"/> type, to the service collection.
+    /// </summary>
+    /// <typeparam name="TAnchor">Anchor type to search</typeparam>
+    /// <param name="services">Service collection</param>
+    /// <param name="lifetime">Lifetime</param>
+    /// <returns>Service collection</returns>
+    public static IServiceCollection AddHandlersFromAssemblyContaining<TAnchor>(this IServiceCollection services,
+        ServiceLifetime lifetime = ServiceLifetime.Scoped)
+    {
+        var definerTypes = typeof(TAnchor).Assembly.ExportedTypes
+            .Where(e => e.GetInterfaces().Where(o => o.IsGenericType).Any(o => o.GetGenericTypeDefinition() == typeof(IHandler<,>)))
+            .Where(e => e is { IsAbstract: false, IsInterface: false })
+            .Select(e => (e, e.GetInterfaces()
+                .Where(o => o.IsGenericType)
+                .Single(o => o.GetGenericTypeDefinition() == typeof(IHandler<,>))));
+
+        foreach (var (handlerType, handlerInterface) in definerTypes)
+        {
+            services.Add(new ServiceDescriptor(handlerInterface, handlerType, lifetime));
         }
 
         return services;
