+++ VSlices.sln
@@ -67,6 +67,18 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VSlices.Domain", "VSlices.D
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VSlices.Domain.UnitTests", "VSlices.Domain.UnitTests\VSlices.Domain.UnitTests.csproj", "{C96835B2-9F94-4834-9C27-31C92772CA45}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{DA82829E-1A46-4F17-B521-53B37313FAD1}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "basic", "basic", "{E68194E7-5DFD-4012-947E-B5702073AA4F}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Domain", "samples\basic\Domain\Domain.csproj", "{C90B89A5-2FB6-4ED2-992F-04265B1D66BE}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CrossCutting", "samples\basic\CrossCutting\CrossCutting.csproj", "{3856C64A-FBEE-4370-862B-83C09565B561}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core", "samples\basic\Core\Core.csproj", "{92F8E539-3C7D-4F13-9AF3-C058FC5BDE3F}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Integration.WebApi", "samples\basic\Integration.WebApi\Integration.WebApi.csproj", "{01EB9DA7-A94D-4950-B8AE-6F1FF5A6BA3A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -161,6 +173,22 @@ Global
 		{C96835B2-9F94-4834-9C27-31C92772CA45}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C96835B2-9F94-4834-9C27-31C92772CA45}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C96835B2-9F94-4834-9C27-31C92772CA45}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C90B89A5-2FB6-4ED2-992F-04265B1D66BE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C90B89A5-2FB6-4ED2-992F-04265B1D66BE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C90B89A5-2FB6-4ED2-992F-04265B1D66BE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C90B89A5-2FB6-4ED2-992F-04265B1D66BE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3856C64A-FBEE-4370-862B-83C09565B561}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3856C64A-FBEE-4370-862B-83C09565B561}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3856C64A-FBEE-4370-862B-83C09565B561}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3856C64A-FBEE-4370-862B-83C09565B561}.Release|Any CPU.Build.0 = Release|Any CPU
+		{92F8E539-3C7D-4F13-9AF3-C058FC5BDE3F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{92F8E539-3C7D-4F13-9AF3-C058FC5BDE3F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{92F8E539-3C7D-4F13-9AF3-C058FC5BDE3F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{92F8E539-3C7D-4F13-9AF3-C058FC5BDE3F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{01EB9DA7-A94D-4950-B8AE-6F1FF5A6BA3A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{01EB9DA7-A94D-4950-B8AE-6F1FF5A6BA3A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{01EB9DA7-A94D-4950-B8AE-6F1FF5A6BA3A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{01EB9DA7-A94D-4950-B8AE-6F1FF5A6BA3A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -196,6 +224,11 @@ Global
 		{6FA91EAA-E1EC-4F56-8E8E-EB49A903C11B} = {671D1A3F-CE83-4CA1-90CD-F42C1F124E28}
 		{B15D0CA4-E28A-4084-8DE8-A740DC35D296} = {E846D8C9-E927-43E8-8B46-CEA3ADB50F6F}
 		{C96835B2-9F94-4834-9C27-31C92772CA45} = {6F484706-EDFC-49EF-A419-D3F6943C2E36}
+		{E68194E7-5DFD-4012-947E-B5702073AA4F} = {DA82829E-1A46-4F17-B521-53B37313FAD1}
+		{C90B89A5-2FB6-4ED2-992F-04265B1D66BE} = {E68194E7-5DFD-4012-947E-B5702073AA4F}
+		{3856C64A-FBEE-4370-862B-83C09565B561} = {E68194E7-5DFD-4012-947E-B5702073AA4F}
+		{92F8E539-3C7D-4F13-9AF3-C058FC5BDE3F} = {E68194E7-5DFD-4012-947E-B5702073AA4F}
+		{01EB9DA7-A94D-4950-B8AE-6F1FF5A6BA3A} = {E68194E7-5DFD-4012-947E-B5702073AA4F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {653AE7A1-A4C4-42F2-9804-FE7E032C3E8C}
